Update Header.cs
As known there are some fields such as CreationTime, AccessTime, ModificationTime in a Lnk file structure. When the "long" data type values in these fields are out of range that the DateTimeOffset function can accept, the Header.cs class throws an exception, and then the process terminates. This fix was developed to this issue.